requirements.txt
instructor
python-dotenv
openai
vllm
logfire
tenacity
